结构调整，代码优化

+++ migration/mysql/20220904193817_create_sys_module_table.sql
@@ -7,7 +7,7 @@ create table sys_module
     `slug`       varchar(20)      not null default '' comment '标示',
     `name`       varchar(20)      not null default '' comment '名称',
     `is_enable`  tinyint unsigned not null default 0 comment '是否启用：1=是；2=否；',
-    `order`      int unsigned     not null default 0 comment '序号',
+    `order`      tinyint unsigned not null default 0 comment '序号',
     `created_at` timestamp        not null default CURRENT_TIMESTAMP,
     `updated_at` timestamp        not null default CURRENT_TIMESTAMP,
     `deleted_at` timestamp                 default null,
